Move CartographerAGoalModal and GoalPublisher into auto cartographer components
