+++ src/ros/nova-gui/nova-gui/src/components/auto/Cartographer/components/GoalPublisher.tsx
